feat(mockotlpserver): a start at module usage (#13)
This refactors the CLI to lib/cli.js, adds a MockOtlpServer export
for module usage. Which 'services' are started (http, grpc, ui)
is now configurable (not exposed to the CLI). The services have
been refactored into Service classes to support exposing an
interface for the service url ('get url()') and shutting down
cleaning ('async close()'). The goal is to support this all
being used in SDK tests.