feat(mockotlpserver): CLI --hostname option (#94)
This allows passing in the hostname that will be used for
the HTTP/gRPC servers to listen on. I added this in case it
is helpful to avoid the 'localhost' default that can be
ambiguously IPv4 or IPv6 depending on the node version. I
don't expect the option to be frequently used.