gather all data (secrets in this case) before doing any parts of the release
The idea here is that we won't create a GitHub release and *then* fail
if the repo is missing a required secret.  We could still have a
partial fail if the secrets are invalid, but there is no perfect
solution for that without transactional code that removes releases
(which would be more maintenance than help).