chore: stop using npm workspaces
THis moves *away* from using npm workspaces. That's a bit of a bummer.
However, AFAICT there is no way to use the package-lock.json when
using npm workspaces to create a dist of just the opentelemetry-node
package. The goal is to have a docker image build of to distro
package, using the versions in package-lock for reproducible
builds. In general a workspaces-package-lock could have a split
of deps at the top-level and in packages/opentelemetry-node/node_modules
that cannot be correctly resolved into a single node_modules
tree without reproducing npm's internal layout logic.