chore(dependabot): drop separate 'monthly' block for updating some npm deps
Today I noticed a new check failure on the latest commit to main. It was to
dependabot.yml. The failure:
> Your .github/dependabot.yml contained invalid details
> Update configs must have a unique combination of 'package-ecosystem', 'directory', and 'target-branch'. Ecosystem 'npm' has overlapping directories.

I don't know if this is a new constraint or was just not noticed before.
I am hoping dropping the 'monthly' block here will suffice. We will
just deal with aws-sdk updates weekly, distracting though that will be.