chore: stop using npm workspaces (#279)
With this change there is now a single "packages/opentelemetry-node/package-lock.json"
file that can unambiguously be used for a reproducible distribution
(i.e. a Docker image or lambda layer zip file).

This moves *away* from using npm workspaces. That's a bit of a bummer.
However, AFAICT there is no way to use the package-lock.json when
using npm workspaces to create a dist of just the opentelemetry-node
package. The goal is to have a docker image build of this
package, using the versions in package-lock for reproducible
builds. In general a workspaces-package-lock could have a split
of deps at the top-level and in packages/opentelemetry-node/node_modules
that cannot be correctly resolved into a single node_modules
tree without reproducing npm's internal layout logic.

Also:
* correct test versionRanges to use caret so '20.0.0' doesn't accidentally pass (20.0.0 >= 18.19.0, oops)
* update dependabot config for the separate packages; deprecate update-otel-deps in the hopes that dependabot can handle otel updates now