chore: convenience 'npm run oneach -- CMD...'
This is a convenience to run a command in each package dir.
Now that we are no longer using npm workspaces, we cannot
'npm run --workspaces ...'. This is mostly similar, except
it is a general command, not an npm script, that is run.
E.g.: npm run oneach -- npm ls @opentelemetry/sdk-node