feat(opentelemetry-node): support for bootstrapping EDOT Node.js with code (#719)
This is as opposed to the zero-code starting via --import of the top-level
entry-point.  This revives the '.../sdk' entry point for exporting APIs
and types required for code bootstrapping. Primarily this is the
startNodeSDK() function -- intentionally a function rather than a
NodeSDK class. Exporting classes creates some compat pain for TypeScript
usage (the types expose internal class properties). My understanding
is that upstream will eventually tend towards an exported *function*
to start the Node SDK.

Also revive typescript-usage.test.js to test that tsc usage works
with a range of TypeScript versions.

Refs: https://github.com/elastic/elastic-otel-node/issues/718
Closes: https://github.com/elastic/elastic-otel-node/issues/700