chore: switch from dependency-check to knip for lint:deps
Recently I had to avoid using the `||=` JS syntax to not break
dependency-check (used for 'npm run lint:deps'). That's lame.
https://github.com/elastic/elastic-otel-node/pull/886/commits/4a77a0ba28717c0ff7161bbbbf8f6bb860c4b1cf

`dependency-check` is deprecated: https://www.npmjs.com/package/dependency-check
We could consider using `knip`. FWIW opentelemetry-js-contrib is using
it.

Here I've chosen to not install knip as a devDep, though we could
consider doing so. I avoided it because `npm init @knip/config`
**installs TypeScript**, which felt a bit overkill to me. We already
have TS at the top-level, so perhaps not bad to add it.