restore a modified version of the old update-otel-deps.js script
We dropped this because dependabot sufficed. However, since adding cooldown in dependabot config
this script is again useful for a quicker update from upstream OTel. Care should be take to
grok the update.

This script is modified from the old one that we had. It now operates on a *single*
package dir. Given this repo no longer uses npm workspaces, the old version didn't
work as is.