feat(opentelemetry-node): include OTEL_ and ELASTIC_OTEL_ envvars in preamble log.info
This includes all likely to be relevant envvars in the preamble log
message to help with clarity and debugging. The *value* of envvars are
included if they are on the allowlist of known to not be sensitive
envvars. Otherwise the value is redacted.

(I think there would still be value to a string representation of the
SDK components. However that is larger work that could be done
separately.)

Closes: #1018