chore(lint): update to eslint@9 (#1221)
Update to eslint@9

Run prettier separate from `eslint` now (as prettier now recommends).
There is a downside to this: prettier does not support the same richness
for "ignore"-comments, so we cannot do the ignoring of a block section
at the top of "packages/opentelemetry-node/lib/instrumentations.js".
If we *really* care about that section, we can move prettier back into
being run with eslint, or look at biome or something else.

I've also changed the `npm run lint` handling a bit:
- `npm run lint` at the top-level will run eslint and prettier for the *whole* repo,
  rather than those being invoked separately for each package. This is a little
  faster. It also still runs all the other `lint:*` scripts in each package
  (via a manually maintained `lint:except-eslint-and-prettier`). Not elegant
  but it works.